@@ -6,9 +6,9 @@ build-backend = "maturin"
 name = "rusterize"
 description = "High performance rasterization tool for Python built in Rust"
 readme = "README.md"
-requires-python = ">=3.10"
+requires-python = ">=3.11"
 license = {file = "LICENSE"}
-keywords = ["fast", "raster"]
+keywords = ["fast", "raster", "geopandas", "xarray"]
 author = {name = "Tommaso Trotto", email = "ttrotto@mail.ubc.ca"}
 classifiers = [
     "License :: OSI Approved :: MIT License",
